restructure tutorial compoents into documentation
