+++ src/features/documentation/concepts.js
@@ -1,0 +1,10 @@
+import React from 'react'
+
+import DocumentationContent from './documentationContent'
+
+const Concepts = () => (
+  <DocumentationContent>
+  </DocumentationContent>
+)
+
+export default Concepts
