+++ src/features/documentation/documentationContent.js
@@ -1,0 +1,40 @@
+import styled from 'styled-components'
+
+import { colour, space, font } from '../common/theme'
+import { media } from '../common/mediaQuery'
+
+const DocumentationContent = styled.section`
+  background-color: ${colour.lightbg};
+  font-size: ${font.large};
+  line-height: ${font.xlarge};
+  padding: ${space.xl} 0 0 10%;
+  font-weight: 300;
+
+  p {
+    width: 60%;
+    margin-bottom: ${space.xl};
+  }
+
+  h3 {
+    color: ${colour.blue};
+    font-size: ${font.xlarge};
+    margin-bottom: ${font.large};
+  }
+
+  button {
+    border: 2px solid ${colour.white};
+    min-width: 100px;
+    color: ${colour.white};
+    border-radius: 2px;
+    background-color: transparent;
+    padding: ${space.m};
+    margin: ${space.m} 0;
+
+    :hover {
+      cursor: pointer;
+      background-color: ${colour.darkbg};
+    }
+  }
+`
+
+export default DocumentationContent
