+++ src/index.js
@@ -10,7 +10,7 @@ import store, { history } from './store'
 import App from './features/app/app'
 import Home from './features/onboarding/home'
 import SignUpContainer from './features/signup/signupContainer'
-import TutorialContainer from './features/tutorial/tutorialContainer'
+import Documentation from './features/documentation/documentation'
 
 
 import 'typeface-lato'
@@ -39,7 +39,7 @@ render(
       <Wrapper>
         <Route exact path='/' component={Home} />
         <Route exact path='/sign-up' component={SignUpContainer} />
-        <Route exact path='/learn' component={TutorialContainer} />
+        <Route path='/learn' component={Documentation} />
         <Route path='/app' component={App} />
       </Wrapper>
     </ConnectedRouter>
