+++ src/reducers/index.js
@@ -5,7 +5,7 @@ import parser from '../features/parser/reducer'
 import simulator from '../features/simulator/reducer'
 import notification from '../features/notifications/reducer'
 import signup from '../features/signup/reducer'
-import tutorial from '../features/tutorial/reducer'
+import documentation from '../features/documentation/reducer'
 
 export default combineReducers({
   routing: routerReducer,
@@ -13,5 +13,5 @@ export default combineReducers({
   simulator,
   notification,
   signup,
-  tutorial
+  documentation
 })
